requirements.txt
asgiref==3.8.1
astroid==3.1.0
attrs==23.2.0
black==24.3.0
click==8.1.7
colorama==0.4.6
coverage==7.4.4
dill==0.3.8
Django==4.2
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
psycopg2==2.9.9
PyJWT==2.8.0
pylint==3.1.0
pytest==8.1.1
pytest-django==4.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
rpds-py==0.18.0
sqlparse==0.4.4
tomlkit==0.12.4
tzdata==2024.1
uritemplate==4.1.1
